Support configuration changes from the shell. syntax: property       => lists the current value of the property property=value => Sets property to value examples: ems=true cycles=5000 core=full
Added "wc" as alias for writeconf and "wl" for writelang(Requested by
DosFreak)
Added support for home/end on the shell(parapente)
Config supports the setting of configuration as well:
config -set "section property=value"

This one should be used for properties that exist in more than one section.
Samplerate for example.

Changed the main shellinput loop. So it checks for a property if the entered
command isn't an internal command or an executable.


git-svn-id: https://dosbox.svn.sourceforge.net/svnroot/dosbox/dosbox/trunk@2190 67dff46e-583c-4055-8ee1-d6b00ecfda6c